elasticsearch/buscador-elastic-noauth-v1/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sentence-transformers==3.0.1
pypdf==4.2.0
beautifulsoup4==4.12.3
markdown==3.6
python-dotenv==1.0.1
requests==2.32.3
